# src/TripBundle/Resources/api/openapi.yaml
openapi: "3.0.5"
info:
    version: 1.0.0
    title: Trips
    license:
        name: GNU AGPLv3
servers:
    -   url: /
security:
    -   BasicAuth: [ ]
paths:
...
    /trip/trips/:
        post:
            summary: Create trip
            operationId: createTrip
            tags:
                - trips
            requestBody:
                description: Data of new trip
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TripCreate'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/Trip'
            responses:
                '200':
                    description: ok
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Trip"
                default:
                    description: unexpected error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DefaultResponse"
        get:
            summary: Get trips
            operationId: getTrips
            tags:
                - trips
            parameters:
                -   in: query
                    name: filter
                    explode: true
                    schema:
                        type: object
                        properties:
                            started_at:
                                type: date
                            finished_at:
                                type: date
                                format: date
                            country:
                                type: string
            responses:
                '200':
                    description: ok
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Trip"
                default:
                    description: unexpected error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DefaultResponse"
    /trip/trips/{tripId}:
        get:
            summary: Get trip information
            operationId: getTrip
            tags:
                - trips
            parameters:
                -   name: tripId
                    in: path
                    required: true
                    schema:
                        type: integer
            responses:
                '200':
                    description: ok
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Trip"
                default:
                    description: unexpected error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DefaultResponse"
        patch:
            summary: Update trip information
            operationId: updateTrip
            tags:
                - trips
            parameters:
                -   name: tripId
                    in: path
                    required: true
                    schema:
                        type: integer
            requestBody:
                description: Updatable data of trip
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TripUpdate'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/Trip'

            responses:
                '200':
                    description: ok
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Trip"
                default:
                    description: unexpected error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DefaultResponse"
        delete:
            summary: Delete trip
            operationId: deleteTrip
            tags:
                - trips
            parameters:
                -   name: tripId
                    in: path
                    required: true
                    schema:
                        type: integer
            responses:
                '200':
                    description: ok
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DefaultResponse"
                default:
                    description: unexpected error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DefaultResponse"
...
components:
    securitySchemes:
        BasicAuth:
            type: http
            scheme: basic
    schemas:
...
        Trip:
            required:
                - id
                - created_by
            allOf:
                -   $ref: '#/components/schemas/TripCreate'
                -   type: object
                    properties:
                        id:
                            type: integer
                            format: int32
                        created_by:
                            type: integer
                            format: int32
        TripCreate:
            required:
                - country
                - started_at
                - finished_at
                - notes
            allOf:
                -   $ref: '#/components/schemas/TripUpdate'
        TripUpdate:
            type: object
            properties:
                country:
                    type: string
                started_at:
                    type: string
                    format: date
                finished_at:
                    type: string
                    format: date
                notes:
                    type: string
...
        DefaultResponse:
            type: object
            required:
                - code
                - message
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string